Ensure that sketches generated by toFreehandSketch(…) are properly added to the blackboard and initialized.
Originally, this function was creating a freehand sketch but not adding it using the addSketch(…) function, meaning that the initialization that happens during addSketch(…) wasn’t happening for sketches that use this function. Sketches that use this include sketches dragged from the glyphs view and sketches that are “unparsed”. This was leading to some errors in the updated experimental ports system I’m developing (which relies on addSketch(…) to do some initialization), and so this needed to be fixed.

I also added some braces around a for loop body for the sake of style.