Allow sketches using primitive types to use them without (un)wrapping
This allows us to go from having to do things like this:
this.defineOutput(“Float”, function() {
	return new AT.Float(value);
}

To this:
this.defineOutput(“Float”, function() {
	return value;
}

For any type (such as Float) that has a toPrimitive function defined.

Same on the input, primitive types have a default preprocessing function defined where toPrimitive is called before returning the value. (Can always access the wrapped value via rawVal or by defining the preprocessing function as the identity function.)