Refactored convert & canConvert to use one code path
With generic types complicating the conversion pipeline, it’s useful to have all the logic relating to conversions in one place. The new _conversionFunction function does just that: it returns the conversion function between two types, or undefined if it doesn’t exist. Thus, convert(…) and canConvert(…) can be implemented in terms of this new function.