Refactored useInputColors to use the new ports system
Turns out that useInputColors didn’t actually work properly at all. It was a completely separate mechanism (deep in sketch.js in modelEndFrame) that was updating the color of the objects. I refactored this whole system: modelEndFrame now uses the sketch’s this.diffuse variable for the color, and useInputColors defines a single input that sets this.diffuse when it receives a value. Specular and ambient never really worked at all, so I’ve removed the code that set those.