Add option for vectors to take in another vector as a control input
Now, if a vector is connected to another of the same axis and dimensions, it uses the input as a control for its own values. If the first input is not of the same axis and dimensions, but is able to be multiplied by the vector, it interprets that input as a value to be multiplied. If a control input is attached, the second input can be used as the multiplication value.