Added conversion from Hz to Function(Seconds, AudioSample)
Ideally this will let us make one of the two speaker sketches obsolete, as they are in the main branch. However, currently the "Hz"-based speaker sketch does some nicer things to the audio signal that prevent discontinuities in the audio signals that can lead to crackling sounds. So we might want to replicate that somewhere in the system.