mesh transformation now correct
TODO: deletion (trying not to reconstruct entire mesh), and also moving between boards. I remember that the 3d mesh object system, which is a node tree, is somewhat separate from the sketches. I can move the mesh sketch to another board, and all the selection logic will work on the other board, but the mesh will stay on the previous node since the node is tied to the sketchPage. Either I figure how to move the node or I do everything on the Unity side.